requirements.txt
absl-py==2.1.0
accelerate==0.21.0
addict==2.4.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
asttokens==2.0.5
async-timeout==4.0.3
attrdict==2.0.1
attrs==24.2.0
beautifulsoup4==4.12.3
bitsandbytes==0.44.1
boto3==1.35.53
botocore==1.35.53
Brotli==1.0.9
bs4==0.0.2
certifi==2024.12.14
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.1.0
cn2an==0.5.22
colorama==0.4.6
comm==0.2.1
contourpy==1.3.0
cpm-kernels==1.0.11
cycler==0.12.1
data-processing==0.1.4
datasets==3.0.2
debugpy==1.8.11
decorator==4.4.2
decord==0.6.0
deepspeed==0.14.4
dill==0.3.8
diskcache==5.6.3
distlib==0.3.9
distro==1.9.0
docker-pycreds==0.4.0
docstring_parser==0.16
einops==0.5.0
einops-exts==0.0.4
et_xmlfile==2.0.0
evaluate==0.4.3
exceptiongroup==1.2.0
executing==0.8.3
fairscale==0.4.13
fastapi==0.103.2
ffmpy==0.4.0
filelock==3.13.1
fire==0.7.0
flash-attn==2.5.5
fonttools==4.54.1
frozenlist==1.5.0
fsspec==2024.9.0
func_timeout==4.3.5
fuzzywuzzy==0.18.0
gitdb==4.0.11
GitPython==3.1.43
gmpy2==2.1.2
gradio==5.1.0
gradio_client==1.4.0
grpcio==1.67.1
h11==0.14.0
h5py==3.12.1
hjson==3.1.0
html2image==2.0.5
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.25.1
human-eval==1.0
identify==2.6.5
idna==3.7
imageio==2.36.0
imageio-ffmpeg==0.5.1
immutabledict==4.2.0
importlib_metadata==8.5.0
interegular==0.3.3
ipdb==0.13.13
ipykernel==6.29.5
ipython==8.30.0
jedi==0.19.2
jieba==0.42.1
Jinja2==3.1.4
jiter==0.6.1
jmespath==1.0.1
joblib==1.4.2
json5==0.9.25
jsonlines==4.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.0
jupyter_core==5.7.2
kiwisolver==1.4.7
langdetect==1.0.9
lark==1.2.2
latex2mathml==3.77.0
Levenshtein==0.26.1
llava==1.2.2.post1
llvmlite==0.43.0
lmdeploy==0.6.2
ltp==4.2.14
ltp-core==0.1.4
ltp-extension==0.1.13
lxml==5.3.0
Markdown==3.7
markdown-it-py==3.0.0
markdown2==2.5.1
MarkupSafe==2.1.3
matplotlib==3.9.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mkl_fft==1.3.11
mkl_random==1.2.8
mkl-service==2.4.0
mmengine==0.10.5
mmengine-lite==0.10.5
moviepy==1.0.3
mpi4py==4.0.1
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
nest-asyncio==1.6.0
networkx==3.2.1
ninja==1.11.1.1
nltk==3.8
nodeenv==1.9.1
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.6.3.3
nvidia-cuda-runtime-cu12==12.6.77
nvidia-curand-cu12==10.3.7.77
nvidia-ml-py==12.560.30
nvidia-nccl-cu12==2.23.4
omegaconf==2.3.0
openai==1.3.5
OpenCC==1.1.9
opencompass==0.3.0
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
openpyxl==3.1.5
orjson==3.10.10
outlines==0.0.46
packaging==24.2
pandas==1.5.3
parso==0.8.4
peft==0.11.1
pexpect==4.8.0
pillow==11.0.0
pip==24.2
platformdirs==3.10.0
portalocker==2.10.1
pre_commit==4.0.1
prettytable==3.12.0
proces==0.1.7
proglog==0.1.10
prompt-toolkit==3.0.43
propcache==0.2.0
protobuf==5.28.3
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyairports==2.1.1
pyarrow==18.0.0
pycountry==24.6.1
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
pyext==0.7
Pygments==2.15.1
pynvml==11.5.3
pyparsing==3.2.0
pypinyin==0.53.0
PySocks==1.7.1
pystack==1.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-Levenshtein==0.26.1
python-multipart==0.0.17
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
qwen-vl-utils==0.0.8
rank-bm25==0.2.2
RapidFuzz==3.10.1
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
retrying==1.3.4
rich==13.9.3
rouge==1.0.1
rouge-chinese==1.0.3
rouge_score==0.1.2
rpds-py==0.20.0
ruff==0.7.1
s3transfer==0.10.3
sacrebleu==2.4.3
safetensors==0.4.5
scikit-learn==1.5.0
scipy==1.14.1
seaborn==0.13.2
semantic-version==2.10.0
sentence-transformers==2.2.2
sentencepiece==0.2.0
sentry-sdk==2.17.0
setproctitle==1.3.3
setuptools==75.1.0
shellingham==1.5.4
shortuuid==1.0.13
shtab==1.7.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
stack-data==0.2.0
starlette==0.27.0
sty==1.0.6
svgwrite==1.4.3
sympy==1.13.3
tabulate==0.9.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
termcolor==2.5.0
threadpoolctl==3.5.0
tiktoken==0.8.0
timeout-decorator==0.5.0
timm==0.6.13
tokenizers==0.15.2
tomli==2.0.2
tomlkit==0.12.0
torch==2.1.2+cu121
torchaudio==2.1.2
torchvision==0.16.2
tornado==6.4.2
tqdm==4.66.6
traitlets==5.14.3
transformers==4.37.2
triton==2.1.0
typeguard==4.4.1
typer==0.12.5
typing_extensions==4.12.2
tyro==0.9.3
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.0
validators==0.34.0
virtualenv==20.29.1
vlmeval==0.1.0
wandb==0.18.5
wavedrom==2.0.3.post3
wcwidth==0.2.5
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.1.3
wheel==0.44.0
XlsxWriter==3.2.0
xxhash==3.5.0
yapf==0.40.2
yarl==1.17.1
zipp==3.20.2
